requirements.txt
streamlit
google-generativeai
python-dotenv
pathlib
youtube_transcript_api